// src/app/api-services/models/jr-register-tenant-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface JrRegisterTenantRequest {
  adminEmail: string;
  adminFirstName: string;
  adminLastName: string;
  adminPassword: string;
  adminUsername: string;
  companyCode: string;
  companyEmail: string;
  companyName: string;
}